% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-fitness-pal.r
\name{get_auth_data}
\alias{get_auth_data}
\title{Retrieves authentication data}
\usage{
get_auth_data()
}
\value{
A list structure containing authentication data.
}
\description{
Retrieves authentication data from MFP, using the session cookie after a succesful login.
}
\author{
Edwin De Jong
}

