% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-fitness-pal.r
\name{get_measurements}
\alias{get_measurements}
\title{Retrieve MFP measurements}
\usage{
get_measurements(context, measurement = "Weight",
  upper_bound = lubridate::now(), lower_bound = upper_bound -
  lubridate::days(30))
}
\arguments{
\item{context}{The context object, containing authentication information.}

\item{measurement}{The type of measurement to retrieve, for example "Weight" or "Waist"}

\item{upper_bound}{The upper bound of measurements as a POSIXct object}

\item{lower_bound}{The lower bound of measurements as a POSIXct object}
}
\value{
A dataframe containing two vectors: "dates" and "measurements"
}
\description{
Retrieves measurements from myfitnesspal.
}
\details{
Retrieves MFP measurements of the given measurement type within the given date interval using
a series of HTTP requests and scraping the measurement from each table.
}
\author{
Edwin De Jong
}

