% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-fitness-pal.r
\name{getContext}
\alias{getContext}
\title{Get context for MFP}
\usage{
getContext(username, password)
}
\arguments{
\item{username}{Username of the account}

\item{password}{The password of the account}
}
\value{
A list containing authentication and user data
}
\description{
Retrieve a context object to retrieve authenticated data from myfitnesspal
}

