% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-fitness-pal.r
\name{mfp_context}
\alias{mfp_context}
\title{Get context for MFP}
\usage{
mfp_context(username, password)
}
\arguments{
\item{username}{Username of the account. Not the email address, but the username found in your profile.}

\item{password}{Password of the account}
}
\value{
List containing authentication and user data
}
\description{
Retrieve a MFP (myfitnesspal) context object to retrieve authenticated data from myfitnesspal.
}

