% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmyfitnesspal.R
\name{mfp_diary}
\alias{mfp_diary}
\title{Retrieve MFP diary}
\usage{
mfp_diary(context, upper_bound = lubridate::now(), lower_bound = upper_bound
  - lubridate::days(300))
}
\arguments{
\item{context}{The context object (can be retrieved using mfp_context)}

\item{upper_bound}{The upper bound of the diary retrieved}

\item{lower_bound}{The lower bound of the diary retrieved}
}
\value{
A data.frame with the diary entries.
}
\description{
Retrieves a MFP diary of foods and exercises.
}
\details{
With this function and an active context, a diary within a certain time-frame can
be downloaded and inserted into a dataframe. The data-frame contains rows with the
date, exercise information and food information.
}
\author{
Edwin De Jong
}

