% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmyfitnesspal.R
\name{mfp_measurements}
\alias{mfp_measurements}
\title{Retrieve MFP measurements}
\usage{
mfp_measurements(context, measurement = "Weight",
  upper_bound = lubridate::now(), lower_bound = upper_bound -
  lubridate::days(30))
}
\arguments{
\item{context}{MFP context object, containing authentication information.}

\item{measurement}{Type of measurement to retrieve, for example "Weight" or "Waist"}

\item{upper_bound}{Upper bound of measurements as a POSIXct object}

\item{lower_bound}{Lower bound of measurements as a POSIXct object}
}
\value{
Dataframe containing two vectors: "dates" and "measurements"
}
\description{
Retrieves measurements of a specific type from myfitnesspal within a given time interval.
}
\details{
Retrieves MFP measurements of the given measurement type within the given date interval using
a series of HTTP requests and scraping the measurement from each table.
}
\author{
Edwin De Jong
}

