% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-fitness-pal.r
\name{get_user_data}
\alias{get_user_data}
\title{Retrieve user data}
\usage{
get_user_data(context)
}
\arguments{
\item{context}{Context object containing authentication data}
}
\value{
List containing user data
}
\description{
Retrieves user data from MyFitnessPall
}
\details{
Given a context object, retrieves user data.
}
\author{
Edwin De Jong
}

